refactor: split monolithic schema into modular files
- Remove src/lib/db/core/schema.ts monolithic file
- Create modular schema files in src/lib/db/schema/:
  - auth.ts - users, sessions, accounts, verifications
  - subscriptions.ts - subscriptionPlans, subscriptions
  - mcp.ts - mcpApiKeys, serviceHealth
  - user-settings.ts - all user settings tables
  - index.ts - re-exports all schema modules
- Create src/lib/db/schema.ts as redirect to schema/index.ts
- Fix broken import paths across codebase
- Update drizzle config to use new schema glob
- Update documentation (architecture.md, development.md)
- Remove obsolete column-factory.ts and its tests
- All 932 tests pass, migrations work correctly
