--- drizzle/0000_init.sql
@@ -1,81 +1,0 @@
-CREATE TABLE `account` (
-	`id` text PRIMARY KEY NOT NULL,
-	`accountId` text NOT NULL,
-	`providerId` text NOT NULL,
-	`userId` text NOT NULL,
-	`accessToken` text,
-	`refreshToken` text,
-	`idToken` text,
-	`accessTokenExpiresAt` integer,
-	`refreshTokenExpiresAt` integer,
-	`scope` text,
-	`password` text,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE TABLE `session` (
-	`id` text PRIMARY KEY NOT NULL,
-	`expiresAt` integer NOT NULL,
-	`token` text NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	`ipAddress` text,
-	`userAgent` text,
-	`userId` text NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE UNIQUE INDEX `session_token_unique` ON `session` (`token`);--> statement-breakpoint
-CREATE TABLE `subscription_plan` (
-	`id` text PRIMARY KEY NOT NULL,
-	`name` text NOT NULL,
-	`description` text,
-	`price` integer NOT NULL,
-	`currency` text DEFAULT 'USD' NOT NULL,
-	`interval` text NOT NULL,
-	`intervalCount` integer DEFAULT 1 NOT NULL,
-	`features` text,
-	`rateLimit` integer NOT NULL,
-	`rateLimitDuration` integer DEFAULT 60000 NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL
-);
---> statement-breakpoint
-CREATE TABLE `subscription` (
-	`id` text PRIMARY KEY NOT NULL,
-	`userId` text NOT NULL,
-	`planId` text NOT NULL,
-	`status` text DEFAULT 'active' NOT NULL,
-	`currentPeriodStart` integer NOT NULL,
-	`currentPeriodEnd` integer NOT NULL,
-	`cancelAtPeriodEnd` integer DEFAULT false NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade,
-	FOREIGN KEY (`planId`) REFERENCES `subscription_plan`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE TABLE `user` (
-	`id` text PRIMARY KEY NOT NULL,
-	`name` text,
-	`email` text NOT NULL,
-	`emailVerified` integer DEFAULT false NOT NULL,
-	`image` text,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	`subscriptionTier` text DEFAULT 'free' NOT NULL,
-	`subscriptionId` text,
-	`subscriptionStatus` text,
-	`subscriptionExpiresAt` integer
-);
---> statement-breakpoint
-CREATE TABLE `verification` (
-	`id` text PRIMARY KEY NOT NULL,
-	`identifier` text NOT NULL,
-	`value` text NOT NULL,
-	`expiresAt` integer NOT NULL,
-	`createdAt` integer,
-	`updatedAt` integer
-);
