--- drizzle/0002_premium_micromax.sql
@@ -1,8 +1,0 @@
-CREATE TABLE `service_health` (
-	`id` integer PRIMARY KEY AUTOINCREMENT NOT NULL,
-	`service_name` text NOT NULL,
-	`status` text NOT NULL,
-	`latency_ms` integer,
-	`error` text,
-	`timestamp` integer DEFAULT CURRENT_TIMESTAMP NOT NULL
-);
