--- drizzle/0003_known_jackal.sql
@@ -1,44 +1,0 @@
-CREATE TABLE `user_settings_account` (
-	`id` text PRIMARY KEY NOT NULL,
-	`userId` text NOT NULL,
-	`name` text DEFAULT '' NOT NULL,
-	`dob` integer,
-	`language` text DEFAULT 'en' NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE TABLE `user_settings_display` (
-	`id` text PRIMARY KEY NOT NULL,
-	`userId` text NOT NULL,
-	`items` text DEFAULT '["recents","home"]' NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE TABLE `user_settings_notifications` (
-	`id` text PRIMARY KEY NOT NULL,
-	`userId` text NOT NULL,
-	`type` text DEFAULT 'all' NOT NULL,
-	`mobile` integer DEFAULT false NOT NULL,
-	`communicationEmails` integer DEFAULT false NOT NULL,
-	`socialEmails` integer DEFAULT true NOT NULL,
-	`marketingEmails` integer DEFAULT false NOT NULL,
-	`securityEmails` integer DEFAULT true NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
---> statement-breakpoint
-CREATE TABLE `user_settings_profile` (
-	`id` text PRIMARY KEY NOT NULL,
-	`userId` text NOT NULL,
-	`username` text NOT NULL,
-	`bio` text DEFAULT '' NOT NULL,
-	`urls` text DEFAULT '[]' NOT NULL,
-	`createdAt` integer NOT NULL,
-	`updatedAt` integer NOT NULL,
-	FOREIGN KEY (`userId`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
-);
