--- drizzle/0004_true_warstar.sql
@@ -1,1 +1,0 @@
-ALTER TABLE `user_settings_profile` ADD `email` text DEFAULT '' NOT NULL;
