+++ drizzle/meta/_journal.json
@@ -5,36 +5,8 @@
     {
       "idx": 0,
       "version": "6",
-      "when": 1774518343314,
-      "tag": "0000_init",
-      "breakpoints": true
-    },
-    {
-      "idx": 1,
-      "version": "6",
-      "when": 1775751613766,
-      "tag": "0001_melted_ego",
-      "breakpoints": true
-    },
-    {
-      "idx": 2,
-      "version": "6",
-      "when": 1776677514138,
-      "tag": "0002_premium_micromax",
-      "breakpoints": true
-    },
-    {
-      "idx": 3,
-      "version": "6",
-      "when": 1776973541173,
-      "tag": "0003_known_jackal",
-      "breakpoints": true
-    },
-    {
-      "idx": 4,
-      "version": "6",
-      "when": 1777130151075,
-      "tag": "0004_true_warstar",
+      "when": 1777304113643,
+      "tag": "0000_ambitious_sunset_bain",
       "breakpoints": true
     }
   ]
