+++ src/lib/db/schema.ts
@@ -1,0 +1,6 @@
+/**
+ * Database schema re-exports.
+ * Redirects to modular schema in ./schema/ directory.
+ */
+
+export * from "./schema/index";
