+++ test/lib/db.test.ts
@@ -3,7 +3,7 @@ import { createClient } from "@libsql/client";
 import { drizzle } from "drizzle-orm/libsql";
 import { eq } from "drizzle-orm";
 import { faker } from "@faker-js/faker";
-import * as schema from "../../src/lib/db/core/schema";
+import * as schema from "../../src/lib/db/schema";
 
 const TEST_DB_PATH = ":memory:";
